add menu and id device

+++ common/adc/adc.h
@@ -2,8 +2,8 @@
 #define _ADC_H
 
 void init_adc1(void);
-uint32_t adc_read_tds_sensor();
-uint32_t adc_read_ph_sensor();
+float adc_read_tds_sensor();
+float adc_read_ph_sensor();
 
 #endif
